:recycle: refactor folder structure
