+++ src/@types/translations/ProjectTranslation.tsx
@@ -1,0 +1,8 @@
+export type ProjectTranslation = {
+  type: TypeTranslation
+}
+
+export type TypeTranslation = {
+  college: string
+  draw: string
+}
