+++ src/app/[lang]/projects/page.tsx
@@ -1,17 +1,8 @@
-import ProjectCard from '@/components/ProjectCard'
+import Projects from '@/features/Projects'
 import getProjectsMetadata from '@/util/functions/getProjectMetadata'
 
-export default function Projects() {
+export default function ProjectsPage() {
   const postMetadata = getProjectsMetadata()
 
-  const postPreviews = postMetadata.map((project) => (
-    <div key={project.id}>
-      <ProjectCard key={project.id} card={project}></ProjectCard>
-    </div>
-  ))
-  return (
-    <div className="flex gap-5 flex-wrap justify-center max-w-project m-auto">
-      {postPreviews}
-    </div>
-  )
+  return <Projects data={postMetadata} />
 }
