+++ src/features/Project/components/ImageSlider/index.tsx
